Split kinematic_constraint_test.cc
